feat: migrate regret storage from AtomicI64 to AtomicI32, REGRET_SCALE 100k->1k
Halves memory usage for regret/strategy/baseline/prediction buffers by
switching from i64 to i32 atomics. REGRET_SCALE reduced from 100,000 to
1,000 to keep values within i32 range while preserving sub-chip precision.

Key changes:
- storage.rs: All 4 buffer types AtomicI64->AtomicI32, REGRET_SCALE 1000,
  regret_floor i32, add_regret uses saturating_add on both paths,
  save/load serialize as i32 with legacy i64 fallback (divide by 100)
- optimizer.rs: CfrOptimizer trait and all 3 impls (DCFR, SAPCFR+, BRCFR+)
  use AtomicI32 slices and cast to i32
- mccfr.rs: prune_threshold i32, delta/strategy_sum casts as i32
  (FullTreeEvTracker/ScenarioEvTracker keep AtomicI64)
- trainer.rs: prune threshold/regret floor computed as i32,
  prev_strategy_sums Vec<i32>, inline LCFR fallback casts as i32
- blueprint_tui_audit.rs: test scale values as i32

+++ crates/core/src/blueprint_v2/mccfr.rs
@@ -648,7 +648,7 @@ pub fn traverse_external(
     traverser: u8,
     node_idx: u32,
     prune: bool,
-    prune_threshold: i64,
+    prune_threshold: i32,
     prune_streets: [bool; 4],
     rng: &mut impl Rng,
     rake_rate: f64,
@@ -848,7 +848,7 @@ pub fn traverse_best_response(
                 if let Some(ps) = predict_storage {
                     for (a, _) in children.iter().enumerate() {
                         let delta = action_values[a] - best;
-                        let delta_scaled = (delta * super::storage::REGRET_SCALE) as i64;
+                        let delta_scaled = (delta * super::storage::REGRET_SCALE) as i32;
                         ps.set_prediction(node_idx, bucket, a, delta_scaled);
                     }
                 }
@@ -885,7 +885,7 @@ fn traverse_traverser(
     children: &[u32],
     num_actions: usize,
     prune: bool,
-    prune_threshold: i64,
+    prune_threshold: i32,
     street: Street,
     prune_streets: [bool; 4],
     rng: &mut impl Rng,
@@ -943,14 +943,14 @@ fn traverse_traverser(
             continue;
         }
         let delta = av - node_value;
-        let delta_scaled = (delta * super::storage::REGRET_SCALE) as i64;
+        let delta_scaled = (delta * super::storage::REGRET_SCALE) as i32;
         storage.add_regret(node_idx, bucket, a, delta_scaled);
         storage.set_prediction(node_idx, bucket, a, delta_scaled);
     }
 
     // Accumulate strategy sums (for computing the average strategy).
     for (a, &s) in strategy.iter().enumerate().take(num_actions) {
-        storage.add_strategy_sum(node_idx, bucket, a, (s * 1000.0) as i64);
+        storage.add_strategy_sum(node_idx, bucket, a, (s * 1000.0) as i32);
     }
 
     // Accumulate EV at tracked scenario nodes.
@@ -1010,7 +1010,7 @@ fn traverse_opponent(
     children: &[u32],
     num_actions: usize,
     prune: bool,
-    prune_threshold: i64,
+    prune_threshold: i32,
     prune_streets: [bool; 4],
     rng: &mut impl Rng,
     rake_rate: f64,
